+++ test/test.apirequest.ts
@@ -1,18 +1,16 @@
-/**
- * Copyright 2018 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2018 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 import {assert} from 'chai';
 import * as crypto from 'crypto';
@@ -207,7 +205,7 @@ describe('createAPIRequest', () => {
       assert(result);
     });
 
-    it('should include directives in the user agent', async () => {
+    it('should include directives in the user agent with local config', async () => {
       const scope = nock(url)
         .get('/')
         .reply(200);
@@ -222,6 +220,57 @@ describe('createAPIRequest', () => {
         requiredParams: [],
         pathParams: [],
         context: fakeContext,
+      });
+      scope.done();
+      // frog/1.0 (jumps) google-api-nodejs-client/0.6.0 (gzip)
+      const userAgent = res.config.headers!['User-Agent'];
+      assert.match(userAgent, /frog\/1.0 \(jumps\)/);
+      assert.match(userAgent, /google-api-nodejs-client\/.* \(gzip\)/);
+    });
+
+    it('should include directives in the user agent with per-service config', async () => {
+      const scope = nock(url)
+        .get('/')
+        .reply(200);
+      const res = await createAPIRequest<FakeParams>({
+        options: {url},
+        params: {},
+        requiredParams: [],
+        pathParams: [],
+        context: {
+          _options: {
+            userAgentDirectives: [
+              {product: 'frog', version: '1.0', comment: 'jumps'},
+            ],
+          },
+        },
+      });
+      scope.done();
+      // frog/1.0 (jumps) google-api-nodejs-client/0.6.0 (gzip)
+      const userAgent = res.config.headers!['User-Agent'];
+      assert.match(userAgent, /frog\/1.0 \(jumps\)/);
+      assert.match(userAgent, /google-api-nodejs-client\/.* \(gzip\)/);
+    });
+
+    it('should include directives in the user agent with global config', async () => {
+      const scope = nock(url)
+        .get('/')
+        .reply(200);
+      const res = await createAPIRequest<FakeParams>({
+        options: {url},
+        params: {},
+        requiredParams: [],
+        pathParams: [],
+        context: {
+          _options: {},
+          google: {
+            _options: {
+              userAgentDirectives: [
+                {product: 'frog', version: '1.0', comment: 'jumps'},
+              ],
+            },
+          },
+        },
       });
       scope.done();
       // frog/1.0 (jumps) google-api-nodejs-client/0.6.0 (gzip)